packages/app/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "foxdata",
    "pyd3graphviz",
    "pyhtmx",
    "utils",
    "Hypercorn == 0.17.3",
    "python-dotenv == 1.1.0",
    "Quart == 0.20.0",
    "uvloop == 0.21.0",
    "XlsxWriter == 3.2.2",
]
